Finally got rid of `parser->token_list`, which means there's no longer a limit to the tokens we can scan. But this require to do the following changed:     * Deleted `_toklst_nospace()` and `_toklst_next()` macros     * Deleted memebers `token_list`, `token_list_len` and `token_list_size` from the mvm_asm_parser_t struct     * Deleted `mvm_asm_token_new()` and implemented its equivalent `mvm_asm_token_yield()` which doesn't require `token_list`     * Deleted `mvm_asm_tokenize()` and implemented its equivalent `mvm_asm_tokenize_next()` (more about it at the end) which calls `mvm_asm_token_yield()` instead of `mvm_asm_token_new()` Changed prefix from MVM_ATS_ to MVM_ATDS_ which is more acurate. It stands for Assembler Tokenizer Data State. Changed `mvm_asm_token_reader_next()` name to `mvm_asm_token_read_next()`
`mvm_asm_tokenize_next()` is status based. Its first argument `tokenizer` is a `tokenizer_t` pointer. The second argument `token` is a `token_t` pointer. Every time you call the function, it will give you the next file token through the second arguemnt and will return `1` if there was a token. If there's no more tokens to give, it returns `0`. And if there was an exception, it return `-1`.